feat: add loadouts command
Add "/loadouts" command for posting a list of randomly generated loadouts.

Uses the API of https://genr8rs.com/Generator/Tf2/LoadoutGenerator to generate loadouts. Data from API is processed in gener8rs_api.py.

One can give an integer as an argument to generate a corresponding amount of numbered loadouts, or leave argument blank for six loadouts with a name for each of the six main players in the server.